backend/requirements.txt
# ---- FASTAPI ----
fastapi==0.119.0
uvicorn[standard]==0.38.0
gunicorn==21.2.0
starlette==0.48.0

# ---- DATABASE ----
SQLAlchemy==2.0.44
psycopg2-binary==2.9.7

# ---- AUTH / SECURITY ----
python-jose==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
email-validator==2.3.0

# ---- UTILS ----
python-dotenv==1.0.1
python-multipart==0.0.20

# ---- AI / GEMINI ----
google-generativeai==0.8.5
google-api-core==2.28.1
protobuf==5.29.5

# ---- COMPATIBILITÉ PYDANTIC V2 ----
pydantic==2.12.3
pydantic_core==2.41.4
typing_extensions==4.15.0

# ---- HTTP CLIENT ----
httpx==0.28.1
